Struct oak_launcher_utils::launcher::Params

pub struct Params {
    pub vmm_binary: PathBuf,
    pub enclave_binary: PathBuf,
    pub app_binary: PathBuf,
    pub bios_binary: PathBuf,
    pub gdb: Option<u16>,
    pub memory_size: Option<String>,
}

Represents parameters used for launching VM instances.

Fields
vmm_binary: PathBuf

Path to the VMM binary to execute.

enclave_binary: PathBuf

Path to the enclave binary to load into the VM.

app_binary: PathBuf

Path to the Oak Functions application binary to be loaded into the enclave.

bios_binary: PathBuf

Path to the BIOS image to use.

gdb: Option<u16>

Port to use for debugging with gdb

memory_size: Option<String>

How much memory to give to the enclave binary, e.g., 256M (M stands for Megabyte, G for Gigabyte).