Trait oak_launcher_utils::launcher::GuestInstance

pub trait GuestInstance {
    // Required methods
    fn wait<'life0, 'async_trait>(
        &'life0 mut self,
    ) -> Pin<Box<dyn Future<Output = Result<ExitStatus>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn kill<'async_trait>(
        self: Box<Self>,
    ) -> Pin<Box<dyn Future<Output = Result<ExitStatus>> + Send + 'async_trait>>
       where Self: 'async_trait;
    fn connect<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = Result<Box<dyn Channel>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

Defines the interface of a launched guest instance. Standardizes the interface of different implementations, e.g. a VM in which the guest is running or the guest running directly as a unix binary.

Required Methods

fn wait<'life0, 'async_trait>( &'life0 mut self, ) -> Pin<Box<dyn Future<Output = Result<ExitStatus>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Wait for the guest instance process to finish.

fn kill<'async_trait>( self: Box<Self>, ) -> Pin<Box<dyn Future<Output = Result<ExitStatus>> + Send + 'async_trait>>
where Self: 'async_trait,

Kill the guest instance.

fn connect<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = Result<Box<dyn Channel>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Creates a channel to communicate with the guest instance.